+++ slack-gamebot/api/endpoints/root_endpoint.rb
@@ -18,6 +18,7 @@ module Api
       mount Api::Endpoints::TeamsEndpoint
       mount Api::Endpoints::GamesEndpoint
       mount Api::Endpoints::SubscriptionsEndpoint
+      mount Api::Endpoints::CreditCardsEndpoint
 
       add_swagger_documentation
     end