+++ spec/api/endpoints/root_endpoint_spec.rb
@@ -7,7 +7,7 @@ describe Api::Endpoints::RootEndpoint do
     get '/api/'
     expect(last_response.status).to eq 200
     links = JSON.parse(last_response.body)['_links']
-    expect(links.keys.sort).to eq(%w(self status team teams user users challenge challenges match matches current_season season seasons subscriptions game games).sort)
+    expect(links.keys.sort).to eq(%w(self status team teams user users challenge challenges credit_cards match matches current_season season seasons subscriptions game games).sort)
   end
   it 'follows all links' do
     get '/api/'
@@ -17,6 +17,7 @@ describe Api::Endpoints::RootEndpoint do
       href = h['href']
       next if href.include?('{') # templated link
       next if href == 'http://example.org/api/subscriptions'
+      next if href == 'http://example.org/api/credit_cards'
       get href.gsub('http://example.org', '')
       expect(last_response.status).to eq 200
       expect(JSON.parse(last_response.body)).to_not eq({})